backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clerk-bot"
version = "0.1.0"
description = "AI-Powered Universal Form Auto-Filler"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "typer[all]>=0.12.0",
    "python-dotenv>=1.0.0",
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "pypdf>=4.0.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "sse-starlette>=1.0.0",
]

[project.scripts]
clerk-bot = "src.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]
